stocktraderapp/requirements.txt
appdirs==1.4.3
click==6.7
dominate==2.3.1
python-dotenv==0.10.1
Flask==0.12.3
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
requests==2.20.0
six==1.10.0
SQLAlchemy==1.1.9
visitor==0.1.3
Werkzeug==0.12.1
WTForms==2.1
